% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework10.R
\name{benchmarks}
\alias{benchmarks}
\title{Laplace density function R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{dlapR} and \code{rwlapR}) and Cpp functions (\code{gibbsC} and \code{vaccC}).
}
\examples{
\dontrun{
tm1 <- microbenchmark::microbenchmark(
  lapR = dlapR(10),
  lapC = dlapC(10)
)
print(summary(tm1)[,c(1,3,5,6)])

tm2 <- microbenchmark::microbenchmark(
  rwR = rwlapR(5000, 2),
  rwC = rwlapC(5000, 2)
)
print(summary(tm2)[,c(1,3,5,6)])
}
}
