% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework07.R
\name{count5test.permutation}
\alias{count5test.permutation}
\title{Comparsion between count5test and permutation version.}
\description{
A Comparsion between permutation test \code{permu.count5} and ordinary count5test \code{count5test} for equal variance based on the maximum number of extreme points that applies when sample sizes are not necessarily equal.
}
\examples{
\dontrun{
# MC-permutation-count 5 test
set.seed(1630)
n1 = c(30, 40, 45, 50, 105)
n2 = c(20, 40, 40, 40, 90)
mu1 = mu2 = 2
sigma1 = sigma2 = 1
m = 1000
permu.compare(group1 = n1, group2 = n2)
}
}
