% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework05.R, R/Homework07.R
\name{count5test}
\alias{count5test}
\title{Count-5-test}
\usage{
count5test(x, y)

count5test(x, y)
}
\arguments{
\item{x, y}{two sample sets}
}
\value{
Test statistic for count-5-test.
}
\description{
The equal variance test by count5test.

ordinary count 5 test.
}
\examples{
\dontrun{
sigma1 = 1
sigma2 = 1.5
m = 1000
n = c(10, 100, 1000)
power_ct = numeric(length(n))
power_f = numeric(length(n))
for(i in 1:length(n)){
  power_ct[i] = mean(replicate(m, expr = {
    x = rnorm(n[i], 0, sigma1)
    y = rnorm(n[i], 0, sigma2)
    count5test(x, y)
  }))
  power_f[i] = mean(replicate(m, expr = {
    x = rnorm(n[i], 0, sigma1)
    y = rnorm(n[i], 0, sigma2)
    var.test(x, y)$p.value < 0.05
  }))
}
list(power_ct = power_ct, power_f = power_f)
}
}
