% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_BeSS_modified.R
\name{CSFGD}
\alias{CSFGD}
\alias{FGD.PDAS}
\title{Competitive Sparse Factor Gradient Descent.}
\usage{
FGD.PDAS(p, X, jt, Data.St, Theta.j = 1, U.Ori, V.Ori, Theta.1st, ActiveSet)
}
\arguments{
\item{p}{nrow of Theta (numeric)}

\item{X}{Data matrix (numeric matrix)}

\item{jt}{item in t (numeric vector)}

\item{Data.St}{Assortment in t (numeric matrix)}

\item{Theta.j}{colwise PDAS (numeric)}

\item{U.Ori}{Initail U (numeric matrix)}

\item{V.Ori}{Initail V (numeric matrix)}

\item{Theta.1st}{Prior estimated Theta using to calculate g,h,delta (numeric matrix)}

\item{ActiveSet}{colwise active set (numeric vector)}
}
\description{
Columnwise PDAS considering data X \code{FGD.PDAS}.

Competitive Sparse Factor Gradient Descent \code{FGD.modBeSS}, which has high speed in calculation.
}
\examples{
\dontrun{
r1 = 2
q1 = 100
p1 = 100
N1 = 200
U1 = matrix(rnorm(p1*2*r1,0,1),p1,2*r1)
V1 = matrix(rnorm(q1*2*r1,0,1),q1,2*r1)
X1 = X.generate(N1, p1)
Theta1 = trueTheta.generate(p1, q1, r1, N1)
ijS1 = ijS.generate(N1, p1, q1, X1, Theta1)
PDAS1 = FGD.PDAS(p = p1, X = X1, jt = ijS1$jt, Data.St = ijS1$St, 
                 U.Ori = U1, V.Ori = V1, Theta.1st = U1\%*\%t(V1), ActiveSet = c(1,3,5,6))
PDAS1
}
}
