% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_BeSS.R
\name{delta.calcu}
\alias{delta.calcu}
\title{Delta Calculation}
\usage{
delta.calcu(g, h, Theta, Theta.j)
}
\arguments{
\item{g, h, Theta, Theta.j}{Are using to calculate Delta.}
}
\description{
Delta Calculation
}
\examples{
\dontrun{
g1 = g.calcu(X1,Theta1,Theta.j=5,ijS1$it,ijS1$jt,ijS1$St)
h1 = h.calcu(X1,Theta1,Theta.j=5,ijS1$it,ijS1$jt,ijS1$St)
delta1 = delta.calcu(g1,h1,Theta1,Theta.j=5)
testgh1 = data.frame(g1 = g1, h1 = h1, delta1 = delta1)
testgh1
}
}
