% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlapC}
\alias{dlapC}
\title{Laplace density function using Rcpp}
\usage{
dlapC(x)
}
\arguments{
\item{x}{the sample points}
}
\value{
the value of density function in x
}
\description{
Laplace density function using Rcpp
}
\examples{
\dontrun{
dlap <- dlapC(5)
dlap
}
}
