% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework10.R
\name{dlapR}
\alias{dlapR}
\title{Laplace density function using R.}
\usage{
dlapR(x)
}
\arguments{
\item{x}{the sample points (numeric)}
}
\value{
the value of density function in x
}
\description{
Laplace density function using R.
}
\examples{
\dontrun{
dla <- dlapR(10)
dla
}
}
