% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework07.R
\name{eqdist.nn}
\alias{eqdist.nn}
\title{\code{NN} test p.value calculation}
\usage{
eqdist.nn(z, sizes, k)
}
\arguments{
\item{z, sizes, k}{using for statistic and p.value calculation}
}
\value{
a list of statistic and p.value
}
\description{
calculate the p.value of \code{NN} test
}
