% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework09.R
\name{eval_f0}
\alias{eval_f0}
\title{eval_f0 Mle}
\usage{
eval_f0(x, x1, n.A = 444, n.B = 132, nOO = 361, nAB = 63)
}
\arguments{
\item{x, x1, n.A, n.B, nOO, nAB}{Mle Constraction}
}
\description{
Mle Calculation
}
