% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_BeSS.R
\name{FGD.BeSS}
\alias{FGD.BeSS}
\title{Factor Gradient Descent.}
\usage{
FGD.BeSS(X, Data.it, Data.jt, Data.St, r, tau, U, V, Actint.m = 100)
}
\arguments{
\item{X}{data matrix (numeric matrix)}

\item{Data.it}{data it (numeric vector)}

\item{Data.jt}{data jt (numeric vector)}

\item{Data.St}{data St (numeric matrix)}

\item{r}{rank of Theta (integer)}

\item{tau}{Tolerance (numeric)}

\item{U}{initail U (numeric matrix)}

\item{V}{initail V (numeric matrix)}

\item{Actint.m}{m max in BeSS (integer)}
}
\value{
the estimate of Theta and error.
}
\description{
Factor Gradient Descent considering data X.
}
\examples{
\dontrun{
FGD.BeSS(X = X1,Data.it = ijS1$it,Data.jt = ijS1$jt,Data.St = ijS1$St,
             r = 2*r1,tau=1e-5,U = U1,V = V1)
}
}
