% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_BeSS_modified.R
\name{FGD.modBeSS}
\alias{FGD.modBeSS}
\title{Competitive Sparse Factor Gradient Descent}
\usage{
FGD.modBeSS(X, Data.it, Data.jt, Data.St, r, U, V, Actint.m = 10, cardi = 5)
}
\arguments{
\item{X}{data matrix (numeric matrix)}

\item{Data.it}{data it (numeric vector)}

\item{Data.jt}{data jt (numeric vector)}

\item{Data.St}{data St (numeric matrix)}

\item{r}{rank of Theta (integer)}

\item{U}{initail U (numeric matrix)}

\item{V}{initail V (numeric matrix)}

\item{Actint.m}{m max in BeSS (integer)}

\item{cardi}{control the level of columnwise sparse (integer)}
}
\description{
Competitive Sparse Factor Gradient Descent considering data X \code{FGD.modBeSS}.
}
\examples{
\dontrun{
r1 = 2
q1 = 100
p1 = 100
N1 = 200
U1 = matrix(rnorm(p1*2*r1,0,1),p1,2*r1)
V1 = matrix(rnorm(q1*2*r1,0,1),q1,2*r1)
X1 = X.generate(N1, p1)
Theta1 = trueTheta.generate(p1, q1, r1, N1)
ijS1 = ijS.generate(N1, p1, q1, X1, Theta1)
Theta.hat1 = FGD.modBeSS(X = X1,Data.it = ijS1$it,Data.jt = ijS1$jt,Data.St = ijS1$St,
                          r = 2*r1,U = U1,V = V1) $ Theta.estimate
RMSE(Theta.hat1, trueTheta)
}
}
