% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_Ordinary.R
\name{FGD.Ordinary}
\alias{FGD.Ordinary}
\title{Factor Gradient Descent.}
\usage{
FGD.Ordinary(X = 0, Data.it, Data.jt, Data.St, r, tau, U, V)
}
\arguments{
\item{X}{data matrix (numeric matrix)}

\item{Data.it}{data it (numeric vector)}

\item{Data.jt}{data jt (numeric vector)}

\item{Data.St}{data St (numeric matrix)}

\item{r}{rank of Theta (integer)}

\item{tau}{Tolerance (numeric)}

\item{U}{initail U (numeric matrix)}

\item{V}{initail V (numeric matrix)}
}
\value{
the estimate of Theta
}
\description{
Factor Gradient Descent considering data X.
}
\examples{
\dontrun{
r1 = 2
q1 = 100
p1 = 100
N1 = 200
U1 = matrix(rnorm(p1*2*r1,0,1),p1,2*r1)
V1 = matrix(rnorm(q1*2*r1,0,1),q1,2*r1)
X1 = X.generate(N1, p1)
Theta1 = trueTheta.generate(p1, q1, r1, N1)
ijS1 = ijS.generate(N1, p1, q1, X1, Theta1)
FGD.Ordinary(X = X1,Data.it = ijS1$it,Data.jt = ijS1$jt,Data.St = ijS1$St,
             r = 2*r1,tau=1e-5,U = U1,V = V1)
}
}
