% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_Ordinary.R
\name{FPR}
\alias{FPR}
\title{FPR, FDR, TPR, TNR, P, R calculation}
\usage{
FPR(Theta.true, Theta.hat)
}
\arguments{
\item{Theta.true}{The true Theta}

\item{Theta.hat}{The estimate of Theta}
}
\description{
FPR, FDR, TPR, TNR, P, R calculation
}
