% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_BeSS.R
\name{g.calcu}
\alias{g.calcu}
\title{Gradient Calculation}
\usage{
g.calcu(X, Theta, Theta.j, it, jt, St)
}
\arguments{
\item{X, Theta, Theta.j, it, jt, St}{Are using to calculate Gradient.}
}
\description{
Gradient Calculation
}
