% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework08.R
\name{GR}
\alias{GR}
\title{Function to compute the diagnnostic statistic R}
\usage{
GR(psi)
}
\arguments{
\item{psi}{the woking matrix used to calculate R.hat}
}
\value{
R.hat
}
\description{
compute the diagnnostic statistic R according to Gelman-Rubin method.
}
