% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_BeSS.R
\name{h.calcu}
\alias{h.calcu}
\title{Hessian Calculation}
\usage{
h.calcu(X, Theta, Theta.j, it, jt, St)
}
\arguments{
\item{X, Theta, Theta.j, it, jt, St}{Are using to calculate diag Hessian.}
}
\description{
Hessian Calculation
}
