% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_Ordinary.R
\name{ijS.generate}
\alias{ijS.generate}
\title{Types, items, assortments generated function}
\usage{
ijS.generate(N, p, q, X, Theta, K = 50)
}
\arguments{
\item{N}{Sample size(time) (numeric)}

\item{p}{nrow of Theta (numeric)}

\item{q}{ncol of Theta (numeric)}

\item{X}{Data Matrix (numeric matrix)}

\item{Theta}{The generated Theta (numeric matrix)}

\item{K}{The maximum number of items can be chosen in time t (numeric)}
}
\description{
Generate types, items, assortments.
}
