% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_BeSS.R
\name{in.St}
\alias{in.St}
\title{in.St}
\usage{
in.St(X, ThetaCol, StRow, itRow, St)
}
\arguments{
\item{X, ThetaCol, StRow, itRow, St}{A preperation for calculating g and h.}
}
\description{
in.St used to calculate g and h
}
