% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework01.R
\name{latticePlot}
\alias{latticePlot}
\title{Different density plots of different sample size.}
\description{
Different density plots of different sample size.
}
\examples{
\dontrun{
n = seq(5, 45, 5)
x = rnorm(sum(n))
y = factor(rep(n, n), labels = paste("n =", n))
lattice::densityplot(~ x | y,
panel = function(x, ...) {
lattice::panel.densityplot(x, col = "DarkOliveGreen", ...)
lattice::panel.mathdensity(dmath = dnorm,
                   args = list(mean = mean(x), sd=sd(x)),
                   col = "darkblue")
})
}
}
