% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_Ordinary.R
\name{loss.multilogit}
\alias{loss.multilogit}
\title{Loss Function with no data matrix.}
\usage{
loss.multilogit(U, V, N, it, jt, St)
}
\arguments{
\item{U}{the estimated U (numeric matrix)}

\item{V}{the estimated V (numeric matrix)}

\item{N}{sample size (integer)}

\item{it}{types (numeric vector)}

\item{jt}{items (numeric vector)}

\item{St}{assortment (numeric matrix)}
}
\value{
the value of loss function
}
\description{
calculate the Loss Function with no data matrix in multi-logit model.
}
\examples{
\dontrun{
Loss.value = loss.multilogit(X = X1, U1, V1, N1, ijS1$it, ijS1$jt, ijS1$St)
Loss.value
}
}
