% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework08.R
\name{mh.lap}
\alias{mh.lap}
\title{M-H sampler}
\usage{
mh.lap(N, sigma, x0 = 0)
}
\arguments{
\item{N}{length of chain}

\item{sigma}{variance of normal distribution}

\item{x0}{initail of x, default value is 0}
}
\description{
M-H sampler
}
\examples{
\dontrun{
# Choose initial values
k = 4
X0 = c(-10, -5, 5, 10)
n = 15000
sigma = 2

# Generate the chains
X = matrix(NA, k, n)
for(i in 1:k) X[i,] = rw.lap(N = n, sigma = sigma, x0 = X0[i])$x

# Compute the diagnostic statistics: mean of ith chain up to time j
psi = t(apply(X, 1, "cumsum"))
for(i in 1:k) psi[i,] = psi[i,] / (1 : ncol(psi))
R.hat = GR(psi)
print(R.hat)
}
\dontrun{
# Plot of Chains psi
par(mfrow = c(2,2))
for(i in 1:k){
  plot(psi[i,1000:n],type = "l",ylab = expression(paste(psi)) )
}
#Plot of R.hat statistics
rhat = numeric(n)
for(j in 501:n) rhat[j] = GR(psi[,1:j])
par(mfrow = c(1,1))
plot(rhat[501:n], type = "l", xlab = " ", ylab = "R")
abline(h = 1.2, lty = 4)
}
}
