% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework05.R
\name{multivariate.skewness.test}
\alias{multivariate.skewness.test}
\title{Calculate the power of Mardia’s multivariate skewness test.}
\description{
the power of the skewness test
}
\examples{
\dontrun{
# power of skewness test
alpha <- 0.1 
n <- 30 
m <- 2500
d <- 2
epsilon <- c(seq(0, .15, .01), seq(.15, 1, .05)) 
N <- length(epsilon) 
pwr <- numeric(N) 
cv <- qchisq(1-alpha,d * (d+1) * (d+2) / 6) 
for (j in 1:N) {
  #for each epsilon
  e <- epsilon[j]
  sktests <- numeric(m)
  for (i in 1:m) {
    #for each replicate
    x1 <- MASS::mvrnorm(n, c(0,0), matrix(c(1,0.3,0.3,1.5), nrow = 2)) 
    x2 <- MASS::mvrnorm(n, c(0,0), matrix(c(100,0.3,0.3,150), nrow = 2)) 
    Iprob = sample(c(1,2), n, replace = TRUE, prob = c(1 - e, e ))
    x = as.integer(Iprob == 1) * x1 + as.integer(Iprob == 2) * x2
    sktests[i] <- as.integer(n*skm(x)/6 >= cv) }
  pwr[j] <- mean(sktests) }
#plot power vs epsilon
plot(epsilon, pwr, type = "b",
     xlab = bquote(epsilon))
abline(h = .1, lty = 3)
se <- sqrt(pwr * (1-pwr) / m) 
#add standard errors
lines(epsilon, pwr+se, lty = 3)
lines(epsilon, pwr-se, lty = 3)
}
}
