% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework07.R
\name{permu.compare}
\alias{permu.compare}
\title{comparsion between two methods}
\usage{
permu.compare(group1, group2)
}
\arguments{
\item{group1, group2}{two groups of samples}
}
\value{
a data.frame for comparision
}
\description{
comparsion between two methods
}
