% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework07.R
\name{permu.count5}
\alias{permu.count5}
\title{permutation-count 5 test for x,y}
\usage{
permu.count5(
  R = 999,
  n1 = 20,
  n2 = 30,
  mu1 = 0,
  mu2 = 0,
  sigma1 = 1,
  sigma2 = 1
)
}
\arguments{
\item{R}{times of replicate}

\item{n1, n2}{two sample sizes}

\item{mu1, mu2}{means of sample sets}

\item{sigma1, sigma2}{true variance of two sample}
}
\description{
permutation version of count 5 test.
}
