% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework08.R
\name{rw.lap}
\alias{rw.lap}
\title{Random Walk Metropolits sampler}
\usage{
rw.lap(N, sigma, x0 = 0)
}
\arguments{
\item{N}{length of chain}

\item{sigma}{variance of normal distribution}

\item{x0}{initail of x, default value is 0}
}
\value{
the generated chain
}
\description{
Random Walk Metropolits sampler for laplace distribution
}
\examples{
\dontrun{
n = 2000
rl1 = rw.lap(N = n, sigma = .05, x0 = 0)
rl2 = rw.lap(N = n, sigma = .5, x0 = 0)
rl3 = rw.lap(N = n, sigma = 2, x0 = 0)
rl4 = rw.lap(N = n, sigma = 16, x0 = 0)
knitr::kable(data.frame(k1 = rl1$k, k2 = rl2$k, k3 = rl3$k, k4 = rl4$k))
par(mfrow=c(2,2))
plot(1:n, rl1$x, ylab = "x", main = expression(paste(sigma, '= 0.05')), type = "l")
plot(1:n, rl2$x, ylab = "x", main = expression(paste(sigma, '= 0.5')), type = "l")
plot(1:n, rl3$x, ylab = "x", main = expression(paste(sigma, '= 2')), type = "l")
plot(1:n, rl4$x, ylab = "x", main = expression(paste(sigma, '= 16')), type = "l")
}
}
