% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwlapC}
\alias{rwlapC}
\title{A Random Walk sampler using Rcpp}
\usage{
rwlapC(N, sigma)
}
\arguments{
\item{N}{the length of chains}

\item{sigma}{the variance of normal distribution}
}
\value{
the generated chain
}
\description{
A Random Walk sampler using Rcpp
}
\examples{
\dontrun{
rwlap <- rwlapC(5000, 2)
plot(1:5000, rwlap$x, ylab = "x", 
main = expression(paste(sigma, '= 2')), type = "l")
}
}
