% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework08.R
\name{S.root}
\alias{S.root}
\title{Root finder}
\usage{
S.root(k)
}
\arguments{
\item{k}{parameter vector k}
}
\value{
unitroot at parameter k
}
\description{
Root finding function
}
\examples{
\dontrun{
set.seed(2156)
k = c(4:25, 100, 500, 1000)
kS = data.frame(k = k , S.root = S.root(k = k))
par(mfrow=c(2,1))
plot(kS$k[1:22], kS$S.root[1:22], type = "l", ylab = "S.root", xlab = "k", lwd = 2, col = "darkred")
plot(kS$k, kS$S.root, type = "l", ylab = "S.root", xlab = "k", lwd = 2, col = "darkred")
knitr::kable(kS[-(3:18),])
}
}
