% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework05.R
\name{skewness.calcul}
\alias{skewness.calcul}
\alias{sk}
\title{skewness calculation}
\usage{
sk(x)
}
\arguments{
\item{x}{Sample data (numeric vector)}
}
\value{
skewness
}
\description{
skewness calculation
}
