% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework05.R
\name{skm}
\alias{skm}
\title{Mardia’s multivariate skewness test}
\usage{
skm(x)
}
\arguments{
\item{x}{Random Matrix}
}
\value{
b1,d
}
\description{
Mardia’s multivariate skewness test and compute type 1 error.
}
\examples{
\dontrun{
m = 100
alpha = 0.05
n = c(10, 100, 1000)
d = 2 
mum = c(0, 0)
sigmam = matrix(c(1, 0.3, 0.3, 2), nrow = 2)
# type 1 error
cvm = qchisq(1 - alpha, d * (d+1) * (d+2) / 6)
p.reject = numeric(length(n))
for(i in 1:length(n)){
  skmtest = numeric(m)
  for(j in 1:m){
    x = MASS::mvrnorm(n[i], mum, sigmam)
    skmtest[j] = as.integer((skm(x) * n[i] / 6 )> cvm)
  }
  p.reject[i] = mean(skmtest)
}
p.reject
}
}
