% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework05.R
\name{sktest}
\alias{sktest}
\title{sktest power calculation}
\usage{
sktest(n, m, cv, type = "norm")
}
\arguments{
\item{n}{Sample size (integer)}

\item{m}{Times of replicate (integer)}

\item{cv}{Crit.values for each n}

\item{type}{The type of distribution.}
}
\value{
mpowerx, power of sktest over m replicate.
}
\description{
sktest power calculation
}
\examples{
\dontrun{
# empirical power
m = 1000
alpha = 0.05
n = c(10, 20, 30, 50, 100, 500) # sample sizes
cv = qnorm(1-alpha, 0, sqrt(6/n)) #crit.values for each n
sknorm = sktest(n = n, m = m,cv = cv, type = "norm")
skbeta = sktest(n = n, m = m,cv = cv, type = "beta")
skt = sktest(n = n, m = m,cv = cv, type = "t")
list(sknorm = sknorm, skbeta = skbeta, skt = skt )
}
}
