% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework07.R
\name{Tn}
\alias{Tn}
\title{\code{NN} test statistic calculation}
\usage{
Tn(z, ix, sizes, k)
}
\arguments{
\item{z, ix, sizes, k}{using for statistic calculation}
}
\description{
calculate the statistic of \code{NN} test
}
