% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_Ordinary.R
\name{trueTheta.generate}
\alias{trueTheta.generate}
\title{True Theta generated function}
\usage{
trueTheta.generate(p, q, r, N)
}
\arguments{
\item{p}{number of row (numeric)}

\item{q}{number of column (numeric)}

\item{r}{rank of Theta (numeric)}

\item{N}{sample sizes (numeric)}
}
\description{
Generate true Theta.
}
