% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGD_Ordinary.R
\name{X.generate}
\alias{X.generate}
\title{Data matrix generated function}
\usage{
X.generate(N, p)
}
\arguments{
\item{N}{sample size (integer)}

\item{p}{dimension of features (integer)}
}
\value{
the generated data matrix X
}
\description{
Generate data X.
}
\examples{
\dontrun{
X.generate(N = 100, p = 10)
}
}
